+++ tests/AbstractQueryBuilderTest.php
@@ -6,14 +6,13 @@ namespace Yiisoft\Db\Tests;
 
 use Closure;
 use JsonException;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\DataProviderExternal;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 use Throwable;
-use Yiisoft\Db\Command\DataType;
+use Yiisoft\Db\Constant\DataType;
 use Yiisoft\Db\Command\Param;
-use Yiisoft\Db\Constant\ColumnType;
-use Yiisoft\Db\Constant\PseudoType;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidArgumentException;
 use Yiisoft\Db\Exception\InvalidConfigException;
@@ -26,8 +25,7 @@ use Yiisoft\Db\Query\QueryInterface;
 use Yiisoft\Db\QueryBuilder\Condition\ArrayOverlapsCondition;
 use Yiisoft\Db\QueryBuilder\Condition\JsonOverlapsCondition;
 use Yiisoft\Db\QueryBuilder\Condition\SimpleCondition;
-use Yiisoft\Db\Schema\Builder\ColumnInterface;
-use Yiisoft\Db\Schema\Column\ColumnSchemaInterface;
+use Yiisoft\Db\Schema\Column\ColumnInterface;
 use Yiisoft\Db\Schema\QuoterInterface;
 use Yiisoft\Db\Tests\Provider\QueryBuilderProvider;
 use Yiisoft\Db\Tests\Support\Assert;
@@ -193,7 +191,7 @@ abstract class AbstractQueryBuilderTest extends TestCase
     }
 
     #[DataProviderExternal(QueryBuilderProvider::class, 'alterColumn')]
-    public function testAlterColumn(string|ColumnSchemaInterface $type, string $expected): void
+    public function testAlterColumn(string|ColumnInterface $type, string $expected): void
     {
         $qb = $this->getConnection()->getQueryBuilder();
 
@@ -225,14 +223,7 @@ abstract class AbstractQueryBuilderTest extends TestCase
         $this->assertSame($expectedParams, $params);
     }
 
-    /**
-     * @dataProvider \Yiisoft\Db\Tests\Provider\QueryBuilderProvider::buildCondition
-     *
-     * @throws Exception
-     * @throws InvalidConfigException
-     * @throws InvalidArgumentException
-     * @throws NotSupportedException
-     */
+    #[DataProviderExternal(QueryBuilderProvider::class, 'buildCondition')]
     public function testBuildCondition(
         array|ExpressionInterface|string $condition,
         string|null $expected,
@@ -453,14 +444,7 @@ abstract class AbstractQueryBuilderTest extends TestCase
         );
     }
 
-    /**
-     * @dataProvider \Yiisoft\Db\Tests\Provider\QueryBuilderProvider::buildLikeCondition
-     *
-     * @throws Exception
-     * @throws InvalidConfigException
-     * @throws InvalidArgumentException
-     * @throws NotSupportedException
-     */
+    #[DataProviderExternal(QueryBuilderProvider::class, 'buildLikeCondition')]
     public function testBuildLikeCondition(
         array|ExpressionInterface $condition,
         string $expected,
@@ -478,7 +462,16 @@ abstract class AbstractQueryBuilderTest extends TestCase
             . (empty($expected) ? '' : ' WHERE ' . DbHelper::replaceQuotes($expected, $db->getDriverName())),
             $sql
         );
-        $this->assertSame($expectedParams, $params);
+        $this->assertSame(array_keys($expectedParams), array_keys($params));
+        foreach ($params as $name => $value) {
+            if ($value instanceof Param) {
+                $this->assertInstanceOf(Param::class, $expectedParams[$name]);
+                $this->assertSame($expectedParams[$name]->getValue(), $value->getValue());
+                $this->assertSame($expectedParams[$name]->getType(), $value->getType());
+            } else {
+                $this->assertSame($expectedParams[$name], $value);
+            }
+        }
     }
 
     public function testBuildLimit(): void
@@ -1784,21 +1777,38 @@ abstract class AbstractQueryBuilderTest extends TestCase
         );
     }
 
-    public function testDropTable(): void
+    public static function dataDropTable(): iterable
+    {
+        yield ['DROP TABLE [[customer]]', null, null];
+        yield ['DROP TABLE IF EXISTS [[customer]]', true, null];
+        yield ['DROP TABLE [[customer]]', false, null];
+        yield ['DROP TABLE [[customer]] CASCADE', null, true];
+        yield ['DROP TABLE [[customer]]', null, false];
+        yield ['DROP TABLE [[customer]]', false, false];
+        yield ['DROP TABLE IF EXISTS [[customer]] CASCADE', true, true];
+        yield ['DROP TABLE IF EXISTS [[customer]]', true, false];
+        yield ['DROP TABLE [[customer]] CASCADE', false, true];
+    }
+
+    #[DataProvider('dataDropTable')]
+    public function testDropTable(string $expected, ?bool $ifExists, ?bool $cascade): void
     {
         $db = $this->getConnection();
-
         $qb = $db->getQueryBuilder();
 
-        $this->assertSame(
-            DbHelper::replaceQuotes(
-                <<<SQL
-                DROP TABLE [[customer]]
-                SQL,
-                $db->getDriverName(),
-            ),
-            $qb->dropTable('customer'),
-        );
+        if ($ifExists === null && $cascade === null) {
+            $sql = $qb->dropTable('customer');
+        } elseif ($ifExists === null) {
+            $sql = $qb->dropTable('customer', cascade: $cascade);
+        } elseif ($cascade === null) {
+            $sql = $qb->dropTable('customer', ifExists: $ifExists);
+        } else {
+            $sql = $qb->dropTable('customer', ifExists: $ifExists, cascade: $cascade);
+        }
+
+        $expectedSql = DbHelper::replaceQuotes($expected, $db->getDriverName());
+
+        $this->assertSame($expectedSql, $sql);
     }
 
     public function testDropUnique(): void
@@ -1833,36 +1843,6 @@ abstract class AbstractQueryBuilderTest extends TestCase
             ),
             $qb->dropview('animal_view'),
         );
-    }
-
-    public function testGetColumnType(): void
-    {
-        $db = $this->getConnection();
-
-        $qb = $db->getQueryBuilder();
-
-        $this->assertSame('pk', $qb->getColumnType(PseudoType::PK));
-        $this->assertSame('upk', $qb->getColumnType(PseudoType::UPK));
-        $this->assertSame('bigpk', $qb->getColumnType(PseudoType::BIGPK));
-        $this->assertSame('ubigpk', $qb->getColumnType(PseudoType::UBIGPK));
-        $this->assertSame('char', $qb->getColumnType(ColumnType::CHAR));
-        $this->assertSame('string', $qb->getColumnType(ColumnType::STRING));
-        $this->assertSame('text', $qb->getColumnType(ColumnType::TEXT));
-        $this->assertSame('tinyint', $qb->getColumnType(ColumnType::TINYINT));
-        $this->assertSame('smallint', $qb->getColumnType(ColumnType::SMALLINT));
-        $this->assertSame('integer', $qb->getColumnType(ColumnType::INTEGER));
-        $this->assertSame('bigint', $qb->getColumnType(ColumnType::BIGINT));
-        $this->assertSame('float', $qb->getColumnType(ColumnType::FLOAT));
-        $this->assertSame('double', $qb->getColumnType(ColumnType::DOUBLE));
-        $this->assertSame('decimal', $qb->getColumnType(ColumnType::DECIMAL));
-        $this->assertSame('datetime', $qb->getColumnType(ColumnType::DATETIME));
-        $this->assertSame('timestamp', $qb->getColumnType(ColumnType::TIMESTAMP));
-        $this->assertSame('time', $qb->getColumnType(ColumnType::TIME));
-        $this->assertSame('date', $qb->getColumnType(ColumnType::DATE));
-        $this->assertSame('binary', $qb->getColumnType(ColumnType::BINARY));
-        $this->assertSame('boolean', $qb->getColumnType(ColumnType::BOOLEAN));
-        $this->assertSame('money', $qb->getColumnType(ColumnType::MONEY));
-        $this->assertSame('json', $qb->getColumnType(ColumnType::JSON));
     }
 
     /**
@@ -1902,7 +1882,7 @@ abstract class AbstractQueryBuilderTest extends TestCase
         $qb = $db->getQueryBuilder();
 
         $this->assertSame($expectedSQL, $qb->insert($table, $columns, $params));
-        $this->assertSame($expectedParams, $params);
+        $this->assertEquals($expectedParams, $params);
     }
 
     /**
@@ -1929,7 +1909,7 @@ abstract class AbstractQueryBuilderTest extends TestCase
 
         $qb = $db->getQueryBuilder();
 
-        $this->assertInstanceOf(QuoterInterface::class, $qb->quoter());
+        $this->assertInstanceOf(QuoterInterface::class, $qb->getQuoter());
     }
 
     public function testRenameColumn(): void
@@ -2269,7 +2249,7 @@ abstract class AbstractQueryBuilderTest extends TestCase
         $qb = $db->getQueryBuilder();
 
         $sql = $qb->update($table, $columns, $condition, $params);
-        $sql = $qb->quoter()->quoteSql($sql);
+        $sql = $db->getQuoter()->quoteSql($sql);
 
         $this->assertSame($expectedSql, $sql);
         $this->assertEquals($expectedParams, $params);
@@ -2397,7 +2377,7 @@ abstract class AbstractQueryBuilderTest extends TestCase
     }
 
     #[DataProviderExternal(QueryBuilderProvider::class, 'buildColumnDefinition')]
-    public function testBuildColumnDefinition(string $expected, ColumnSchemaInterface|string $column): void
+    public function testBuildColumnDefinition(string $expected, ColumnInterface|string $column): void
     {
         $db = $this->getConnection();
         $qb = $db->getQueryBuilder();