+++ tests/Support/Stub/ColumnDefinitionBuilder.php
@@ -6,7 +6,7 @@ namespace Yiisoft\Db\Tests\Support\Stub;
 
 use Yiisoft\Db\Constant\ColumnType;
 use Yiisoft\Db\QueryBuilder\AbstractColumnDefinitionBuilder;
-use Yiisoft\Db\Schema\Column\ColumnSchemaInterface;
+use Yiisoft\Db\Schema\Column\ColumnInterface;
 
 final class ColumnDefinitionBuilder extends AbstractColumnDefinitionBuilder
 {
@@ -36,7 +36,24 @@ final class ColumnDefinitionBuilder extends AbstractColumnDefinitionBuilder
         'decimal',
     ];
 
-    protected function getDbType(ColumnSchemaInterface $column): string
+    protected function buildCheck(ColumnInterface $column): string
+    {
+        $check = $column->getCheck();
+
+        if (empty($check)) {
+            $columnName = $column->getName();
+
+            if (!empty($columnName) && $column->getType() === ColumnType::JSON) {
+                return ' CHECK (json_valid(' . $this->queryBuilder->getQuoter()->quoteColumnName($columnName) . '))';
+            }
+
+            return '';
+        }
+
+        return " CHECK ($check)";
+    }
+
+    protected function getDbType(ColumnInterface $column): string
     {
         return $column->getDbType() ?? match ($column->getType()) {
             ColumnType::BOOLEAN => 'boolean',